school  upd

+++ printf/libft/ft_itoa_base.c
@@ -3,54 +3,54 @@
 /*                                                        :::      ::::::::   */
 /*   ft_itoa_base.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lmeribal <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: lmeribal <lmeribal@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/17 13:48:49 by lmeribal          #+#    #+#             */
-/*   Updated: 2021/01/17 22:00:45 by lmeribal         ###   ########.fr       */
+/*   Updated: 2021/01/18 21:26:00 by lmeribal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int ft_power(unsigned int nb, int pow)
+unsigned long int ft_power(unsigned long int nb, int pow)
 {
-    unsigned int res;
+	unsigned long int res;
 
-    res = 1;
-    if (pow == 0)
-        return (1);
-    while (pow)
-    {
-        res = res * nb;
-        pow--;
-    }
-    return (res);
+	res = 1;
+	if (pow == 0)
+		return (1);
+	while (pow)
+	{
+		res = res * nb;
+		pow--;
+	}
+	return (res);
 }
 
-char *ft_itoa_base(unsigned int value, int base)
+char *ft_itoa_base(unsigned long int value, int base)
 {
-    int i;
-    int neg;
-    char *nbr;
+	int i;
+	int neg;
+	char *nbr;
 
-    i = 1;
-    neg = 0;
-    if (value < 0)
-    {
-        if (base == 10)
-            neg = 1;
-        value *= -1;
-    }
-    while (ft_power(base, i) - 1 < value)
-        i++;
-    nbr = (char*)malloc(sizeof(nbr) * i);
-    nbr[i + neg] = '\0';
-    while (i-- > 0)
-    {
-        nbr[i + neg] = (value % base) + (value % base > 9 ? 'a' - 10 : '0');
-        value = value / base;
-    }
-    if (neg)
-        nbr[0] = '-';
-    return (nbr);
+	i = 1;
+	neg = 0;
+	if (value < 0)
+	{
+		if (base == 10)
+			neg = 1;
+		value *= -1;
+	}
+	while (ft_power(base, i) - 1 < value)
+		i++;
+	nbr = (char*)malloc(sizeof(nbr) * i);
+	nbr[i + neg] = '\0';
+	while (i-- > 0)
+	{
+		nbr[i + neg] = (value % base) + (value % base > 9 ? 'a' - 10 : '0');
+		value = value / base;
+	}
+	if (neg)
+		nbr[0] = '-';
+	return (nbr);
 }
